Work in progress : working on submitting results

+++ SG_Automation/SeleniumSandbox.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python -u
 
 import base64
+import datetime
 import glob
 import hashlib
 import json
@@ -10,6 +11,7 @@ import shutil
 import signal
 import subprocess
 import sys
+import time
 import urllib2
 import urlparse
 
@@ -17,6 +19,10 @@ import testrail
 
 from optparse import OptionParser
 from pprint import pprint as pp
+
+def ppjson(doc):
+    # print json.dumps(doc, sort_keys=True, indent=4, separators=(',', ': '))
+    pp(doc, indent=4, width=1)
 
 def get_site_version(url):
     try:
@@ -50,6 +56,12 @@ class GitHubError(Exception):
         return str(self.message)
 
 class TestRailError(Exception):
+    def __init__(self, message):
+        self.message = message
+    def __str__(self):
+        return str(self.message)
+
+class TestRailRunInvalid(Exception):
     def __init__(self, message):
         self.message = message
     def __str__(self):
@@ -89,6 +101,8 @@ class SeleniumSandbox:
             self.testrail_runs = self.get_testrail_runs()
             self.testrail_plans = self.get_testrail_plans()
         self.debugging = debugging
+        self.testrail_available_cases = {}
+        self.target_url = None
         self.user = self.get_elems("https://api.github.com/user")
 
     def set_work_folder(self, work_folder):
@@ -117,14 +131,14 @@ class SeleniumSandbox:
         runs = {}
         if self.testrail_project_id:
             for run in self.testrail.send_get('get_runs/%d&is_completed=0' % self.testrail_project_id):
-                runs[str(run['id'])] = run
+                runs[run['id']] = run
         return runs
 
     def get_testrail_plans(self):
         plans = {}
         if self.testrail_project_id:
             for plan in self.testrail.send_get('get_plans/%d&is_completed=0' % self.testrail_project_id):
-                plans[str(plan['id'])] = plan
+                plans[plan['id']] = plan
         return plans
 
     def fetch_shotgun_files(self, shotgun_version):
@@ -236,6 +250,7 @@ class SeleniumSandbox:
     def sync_filesystem(self, sha=None, prefix=None):
         if sha is None:
             sha = self.selenium_version
+            self.testrail_available_cases = {}
 
         tree = self.get_tree(sha)
 
@@ -250,10 +265,13 @@ class SeleniumSandbox:
         diff = [x for x in actual if x not in expected]
         for i in diff:
             obj = prefix + i
+            # Do not delete the build folder
             if os.path.isdir(obj):
-                print("Removing folder: %s" % obj)
-                shutil.rmtree(obj)
-            else:
+                if os.path.join(self.work_folder, "suites", "build") != obj:
+                    print("Removing folder: %s" % obj)
+                    shutil.rmtree(obj)
+            # do not delete the config.xml file
+            elif os.path.join(self.work_folder, "suites", "config", "config.xml") != obj:
                 print("Removing file: %s" % obj)
                 os.remove(obj)
 
@@ -262,6 +280,11 @@ class SeleniumSandbox:
             path = prefix + elem["path"]
             sha = elem["sha"]
             if elem["type"] == "tree":
+                is_case = re.match('^C([0-9]+)$', elem["path"])
+                # If it is a folder matching a test case ID, and it is in test_rail
+                # @FIXME: this is a very brittle test...
+                if is_case and path.startswith(os.path.join(self.work_folder, "suites", "test_rail", "")):
+                    self.testrail_available_cases[int(is_case.group(1))] = path
                 if not os.path.exists(path):
                     print("creating folder: " + path)
                     os.mkdir(path)
@@ -273,6 +296,9 @@ class SeleniumSandbox:
                     if os.path.exists(path):
                         statInfo = os.stat(path)
                         fileSha1 = hashlib.sha1("blob " + str(statInfo.st_size) + "\0" + open(path, "rb").read()).hexdigest()
+                        # @FIXME: DO NOT SUBMIT THE FOLLOWING TWO LINES !!! WORK CODE ONLY
+                        if os.path.join("suites", "runTest.command") in path:
+                           continue
                         if fileSha1 != sha:
                             print("updating file: " + path)
                             shutil.copy(self.get_git_object_file(sha), path)
@@ -294,6 +320,11 @@ class SeleniumSandbox:
 
     def generate_config(self, options):
         configFolder = os.path.join(self.work_folder, "suites", "config")
+        if 'sg_config__url' in options:
+            self.target_url = options['sg_config__url']
+        else:
+            self.target_url = None
+
         if os.path.exists(configFolder):
             configFile = open(os.path.join(configFolder, "config.xml"), "w")
             configFile.write(
@@ -312,7 +343,7 @@ class SeleniumSandbox:
             raise UnsupportedBranch("This site does not support Selenium tests")
 
 
-    def run_tests(self, test_suite):
+    def execute_suite(self, test_suite):
         if not test_suite.endswith(self.command_file):
             test_suite = os.path.join(test_suite, self.command_file)
         if not test_suite.startswith(self.work_folder):
@@ -332,6 +363,112 @@ class SeleniumSandbox:
         else:
             raise SuiteNotFound("Non existent test suite %s" % test_suite)
 
+    def get_testrail_runs_from_plan(self, test_plan):
+        runs = []
+        if test_plan not in self.testrail_plans:
+            raise TestRailRunInvalid('TestRail plan %s does not exist' % test_plan)
+        plan = self.testrail.send_get('get_plan/%s' % test_plan)
+        for entry in plan['entries']:
+            for run in entry['runs']:
+                # @FIXME: Okay... not to sure about that... perhaps a separate data member...
+                self.testrail_runs[run['id']] = run
+                runs.append(run['id'])
+        return runs
+
+    def execute_run(self, test_run):
+        targets = []
+        tests = {}
+        if test_run in self.testrail_runs:
+            targets = [test_run]
+        elif test_run in self.testrail_plans:
+            targets = self.get_testrail_runs_from_plan(test_run)
+        else:
+            raise TestRailRunInvalid('TestRail run or plan %s does not exist' % test_run)
+
+        for target in targets:
+            tests[target] = {}
+            for test in self.testrail.send_get('get_tests/%s' % target):
+                if test['case_id'] not in self.testrail_available_cases:
+                    print "WARNING: skipping test case %s (%s) as it is not present in this codebase/version" % (
+                        test['case_id'], test['title']
+                    )
+                else:
+                    # tests[target][test['id']] = test['case_id']
+                    tests[target][test['id']] = test
+        # ppjson(targets)
+        # print "=========================="
+        # ppjson(tests)
+        # print "=========================="
+        # ppjson(tests[3399])
+        # print "=========================="
+        # ppjson(tests[3399][150071])
+        # print "=========================="
+
+        # if not test_suite.endswith(self.command_file):
+        #     test_suite = os.path.join(test_suite, self.command_file)
+        # if not test_suite.startswith(self.work_folder):
+        #     test_suite = os.path.join(self.work_folder, test_suite)
+        targets.sort()
+        for target in targets:
+            print("INFO: Running TestRail test run: %s - %s" % (target, self.testrail_runs[target]['name']))
+            # ppjson(tests[target])
+            keys = tests[target].keys()
+            keys.sort()
+            netloc = ""
+            if self.target_url:
+                netloc = urlparse.urlparse(self.target_url).netloc
+
+            build_folder = os.path.join(
+                self.get_work_folder(),
+                "suites",
+                "build",
+                netloc,
+                datetime.datetime.now().strftime("%Y-%m-%d_%Hh%Mm%Ss"))
+            # if not os.path.exists(build_folder):
+            #     print "Build results will be located in folder %s" % build_folder
+                # os.makedirs(build_folder)
+            # print "------> %s" % build_folder
+            results = {"results": []}
+            for test in keys:
+                # print "%s " % type(test),
+                # print "target: %d (%s)" % (target, type(target))
+                # print "test: %d (%s)" % (test, type(test))
+                # ppjson(tests[target][test]['case_id'])
+                test_suite = self.testrail_available_cases[tests[target][test]['case_id']]
+                test_suite = os.path.join(test_suite, self.command_file)
+                # print '%s - %s' % (test, test_suite)
+                if os.path.exists(test_suite):
+                    startTime = time.time()
+                    stopTime = None
+                    self.subProc = subprocess.Popen(test_suite, env={"BUILD_FOLDER": build_folder})
+                    code = -1
+                    try:
+                        code = self.subProc.wait()
+                        stopTime = time.time()
+                    except (KeyboardInterrupt, SystemExit):
+                        self.subProc.kill()
+                    except Exception as e:
+                        self.subProc.kill()
+                        raise
+                    if code == 0 or code == 1 or code == -1:
+                        status_id = 5
+                        result = {
+                            'test_id': 150788,
+                            'status_id': status_id,
+                            'comment': 'from Automation on https://6-3-automation.shotgunstudio.com',
+                            'elapsed': '345s',
+                            'custom_os': [10],
+                            'custom_webbrowser': [20],
+                            'version': 'v6.3.0-dev (build cb789f8)'
+                        }
+                    elif code == -1:
+                        print "-----------------------------------> ABORT ABORT ABORT"
+                    else:
+                        print "-----------------------------------> WTF!!!!!!!!!: %d" % code
+                    # return code
+                else:
+                    raise SuiteNotFound("Non existent test suite %s" % test_suite)
+            ppjson(results)
 
     def signal_handler(self, sig, frm):
         sys.exit(1)
@@ -374,6 +511,14 @@ def main(argv):
         print "Options --suite and --testrail-run are mutually exclusive."
         parser.print_help()
         sys.exit(2)
+
+    if options.testrail_run:
+        if re.match('^[1-9][0-9]*$', options.testrail_run):
+            options.testrail_run = int(options.testrail_run)
+        else:
+            print "A testrail-run argument must be a number."
+            parser.print_help()
+            sys.exit(2)
 
     if options.testrail_run and not options.testrail_token:
         print "A testrail-run argument requires a valid TestRail token."
@@ -460,10 +605,13 @@ def main(argv):
 
         if options.suite:
             print("INFO: Running tests from %s" % options.suite)
-            return_value = sandbox.run_tests(options.suite)
+            return_value = sandbox.execute_suite(options.suite)
             print("INFO:     Test completed")
         elif options.testrail_run:
-            pass
+            if options.testrail_run in sandbox.testrail_plans:
+                print("INFO: Running TestRail test plan: %s - %s" % (options.testrail_run, sandbox.testrail_plans[options.testrail_run]['name']))
+            return_value = sandbox.execute_run(options.testrail_run)
+            print("INFO:     Test completed")
     except UnsupportedBranch as e:
         print("ERROR: This Shotgun site does not support Selenium automation!")
         sys.exit(2)
